ml-service/pyproject.toml
[tool.poetry]
name = "ml-service"
version = "0.1.0"
description = "Matobev ML service for video analysis"
authors = ["Devin AI <devin-ai-integration[bot]@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = "^0.30.6"
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"
httpx = "^0.27.2"
Pillow = "^10.4.0"
opencv-python-headless = "^4.10.0.84"
numpy = "^2.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
